refactor: add constraints to entity, add equals and hashCode to entity and item
