+++ src/main/java/ru/snsin/cakefactory/storage/CakeEntity.java
@@ -2,13 +2,17 @@ package ru.snsin.cakefactory.storage;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 import java.math.BigDecimal;
 
 @Entity
 public class CakeEntity implements Serializable {
     private String id;
+    @NotBlank
     private String name;
+    @NotNull
     private BigDecimal price;
 
     public void setId(String id) {
@@ -16,23 +20,38 @@ public class CakeEntity implements Serializable {
     }
 
     @Id
-    public String getId() {
+    String getId() {
         return id;
     }
 
-    public String getName() {
+    String getName() {
         return name;
     }
 
-    public void setName(String name) {
+    void setName(String name) {
         this.name = name;
     }
 
-    public BigDecimal getPrice() {
+    BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    void setPrice(BigDecimal price) {
         this.price = price;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        CakeEntity that = (CakeEntity) o;
+
+        return id.equals(that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return id.hashCode();
     }
 }
