--- src/test/java/ru/snsin/cakefactory/CakeFactoryApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.snsin.cakefactory;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CakeFactoryApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
