+++ src/test/java/ru/snsin/cakefactory/services/CakeCatalogServiceImplTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import ru.snsin.cakefactory.domain.CakeItem;
+import ru.snsin.cakefactory.storage.CakeCatalogServiceImpl;
 import ru.snsin.cakefactory.storage.CakeRepository;
 
 import java.math.BigDecimal;
